requirements.txt
aiohttp==3.6.2
aiohttp-jinja2==1.2.0
antlr4-python3-runtime==4.7.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.1
certifi==2019.11.28
chardet==3.0.4
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==5.2.1
html2text==2019.9.26
idna==2.8
jieba3k==0.35.1
Jinja2==2.10.3
joblib==0.14.1
lxml==4.4.2
MarkupSafe==1.1.1
multidict==4.7.1
newspaper3k==0.2.8
nltk==3.4.5
numpy==1.17.4
pandas==0.25.3
Pillow==6.2.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
requests==2.22.0
requests-file==1.4.3
scikit-learn==0.22
scipy==1.3.3
simplejson==3.17.0
six==1.13.0
sklearn==0.0
soupsieve==1.9.5
stix2==1.2.1
stix2-patterns==1.2.1
taxii2-client==0.5.0
tinysegmenter==0.3
tldextract==2.2.2
urllib3==1.25.7
yarl==1.4.2
